+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 
 ### Changed
 - cats: fix issue where `startfile` field gets wrongly updated [PR #1346]
+- Python Plugins: Avoid pop(0) performance impact [PR #1351]
 
 [PR #1335]: https://github.com/bareos/bareos/pull/1335
 [PR #1346]: https://github.com/bareos/bareos/pull/1346
+[PR #1351]: https://github.com/bareos/bareos/pull/1351
 [unreleased]: https://github.com/bareos/bareos/tree/master